requirements.txt
# Gumbel AlphaZero for Chinese Chess - Dependencies

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
h5py>=3.8.0
pandas>=2.0.0

# Configuration
pyyaml>=6.0

# Progress Bars
tqdm>=4.66.0

# Logging and Visualization
tensorboard>=2.13.0

# Testing
pytest>=7.0.0

# Utilities
python-dateutil>=2.8.0
